+++ sugarloaf/src/font/constants.rs
@@ -5,58 +5,57 @@ macro_rules! font {
     };
 }
 
-pub const DEFAULT_FONT_FAMILY: &str = "cascadiamono";
-pub const DEFAULT_FONT_FAMILY_VARIANT: &str = "cascadiacode";
+pub const DEFAULT_FONT_FAMILY: &str = "cascadiacode";
 
 // Fonts:
-// CascadiaMonoPL-Bold.ttf
-// CascadiaMonoPL-BoldItalic.ttf
-// CascadiaMonoPL-ExtraLight.ttf
-// CascadiaMonoPL-ExtraLightItalic.ttf
-// CascadiaMonoPL-Italic.ttf
-// CascadiaMonoPL-Light.ttf
-// CascadiaMonoPL-LightItalic.ttf
-// CascadiaMonoPL-Regular.ttf
-// CascadiaMonoPL-SemiBold.ttf
-// CascadiaMonoPL-SemiBoldItalic.ttf
-// CascadiaMonoPL-SemiLight.ttf
-// CascadiaMonoPL-SemiLightItalic.ttf
+// CascadiaCodePL-Bold.ttf
+// CascadiaCodePL-BoldItalic.ttf
+// CascadiaCodePL-ExtraLight.ttf
+// CascadiaCodePL-ExtraLightItalic.ttf
+// CascadiaCodePL-Italic.ttf
+// CascadiaCodePL-Light.ttf
+// CascadiaCodePL-LightItalic.ttf
+// CascadiaCodePL-Regular.ttf
+// CascadiaCodePL-SemiBold.ttf
+// CascadiaCodePL-SemiBoldItalic.ttf
+// CascadiaCodePL-SemiLight.ttf
+// CascadiaCodePL-SemiLightItalic.ttf
 
 pub const FONT_CASCADIAMONO_BOLD: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-Bold.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-Bold.otf");
 
 pub const FONT_CASCADIAMONO_BOLD_ITALIC: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-BoldItalic.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-BoldItalic.otf");
 
 pub const FONT_CASCADIAMONO_EXTRA_LIGHT: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-ExtraLight.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-ExtraLight.otf");
 
 pub const FONT_CASCADIAMONO_EXTRA_LIGHT_ITALIC: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-ExtraLightItalic.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-ExtraLightItalic.otf");
 
 pub const FONT_CASCADIAMONO_ITALIC: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-Italic.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-Italic.otf");
 
 pub const FONT_CASCADIAMONO_LIGHT: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-Light.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-Light.otf");
 
 pub const FONT_CASCADIAMONO_LIGHT_ITALIC: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-LightItalic.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-LightItalic.otf");
 
 pub const FONT_CASCADIAMONO_REGULAR: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-Regular.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-Regular.otf");
 
 pub const FONT_CASCADIAMONO_SEMI_BOLD: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-SemiBold.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-SemiBold.otf");
 
 pub const FONT_CASCADIAMONO_SEMI_BOLD_ITALIC: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-SemiBoldItalic.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-SemiBoldItalic.otf");
 
 pub const FONT_CASCADIAMONO_SEMI_LIGHT: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-SemiLight.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-SemiLight.otf");
 
 pub const FONT_CASCADIAMONO_SEMI_LIGHT_ITALIC: &[u8] =
-    font!("./resources/CascadiaMono/CascadiaMonoPL-SemiLightItalic.otf");
+    font!("./resources/CascadiaCode/CascadiaCodePL-SemiLightItalic.otf");
 
 pub const FONT_SYMBOLS_NERD_FONT_MONO: &[u8] =
     font!("./resources/SymbolsNerdFontMono/SymbolsNerdFontMono-Regular.ttf");